requirements.txt
mysql-connector-python
pandas
bokeh
numpy
scipy
selenium
psycopg2-binary
